+++ pact-jvm-consumer/src/main/kotlin/au/com/dius/pact/consumer/PactConsumerConfig.kt
@@ -1,0 +1,5 @@
+package au.com.dius.pact.consumer
+
+object PactConsumerConfig {
+  val pactDirectory: String = System.getProperty("pact.rootDir", "target/pacts")
+}